requirements.txt
django
psycopg2
Jinja2
djangorestframework
